app/requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-solo==2.1.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot==13.4.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
tzlocal==5.0.1
urllib3==2.0.2
psycopg2-binary==2.9.3
django-object-actions==4.1.0
gunicorn==20.1.0
